requirements/requirements.dev.txt
asgiref==3.6.0
autopep8==2.0.1
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coverage==7.1.0
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.6
django-allauth==0.52.0
django-autoslug==1.9.8
django-cors-headers==3.13.0
django-filter==23.5
django-ipware==4.0.2
django-modeltranslation==0.18.9
django-recaptcha==3.0.0
django-summernote==0.8.20.0
djangorestframework==3.14.0
drf-recaptcha==2.2.2
flake8==6.0.0
idna==3.4
Markdown==3.4.1
mccabe==0.7.0
oauthlib==3.2.2
Pillow==9.4.0
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
webencodings==0.5.1
